+++ ai-prompt-template/src/applicator.rs
