+++ ai-prompt-template/src/openai.rs
@@ -1,3 +1,4 @@
+// Copyright 2023 Salesforce, Inc. All rights reserved.
 use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
